+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [2.0.0] - 2019-10-17
+
+- do not decode workspace variable values - expect strings
+- drop support for workspace Terraform HCL variables
+
 ## [1.2.1] - 2019-10-16
 
 - decode Terraform HCL workspace variable values as `JSON`
@@ -21,8 +26,9 @@ All notable changes to this project will be documented in this file.
 
 - Initial release
 
-[Unreleased]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v1.2.1...HEAD
-[1.2.0]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v1.2.0...v1.2.1
+[Unreleased]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v1.2.1...v2.0.0
+[1.2.1]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/nephosolutions/terraform-google-gcp-project/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/nephosolutions/terraform-google-gcp-project/releases/tag/v1.0.0